Separate EmojiText component in native and web variant with custom implementation
